commit with extra wrapper component
