+++ src/app/shared/form/builder/ds-dynamic-form-ui/ds-dynamic-form-control-container.component.html
@@ -1,6 +1,4 @@
-<div [class.form-group]="(model.type !== 'GROUP' && asBootstrapFormGroup) || getClass('element', 'container').includes('form-group')"
-     [formGroup]="group"
-     [ngClass]="[getClass('element', 'container'), getClass('grid', 'container')]">
+
 
     <label *ngIf="!isCheckbox && hasLabel"
            [for]="model.id"
@@ -9,7 +7,7 @@
 
     <ng-container *ngTemplateOutlet="startTemplate?.templateRef; context: model"></ng-container>
 
-    <div [ngClass]="{'form-row': model.hasLanguages || hasRelationLookup }">
+    <div [ngClass]="{'form-row': model.hasLanguages }">
         <div [ngClass]="getClass('grid', 'control')">
 
             <ng-container #componentViewContainer></ng-container>
@@ -36,32 +34,8 @@
                 <option *ngFor="let lang of model.languageCodes" [value]="lang.code">{{lang.display}}</option>
             </select>
         </div>
-
-        <div *ngIf="hasRelationLookup" class="col-auto text-center">
-            <button class="btn btn-secondary"
-                    type="submit"
-                    ngbTooltip="{{'form.lookup-help' | translate}}"
-                    placement="top"
-                    (click)="openLookup(); $event.stopPropagation();">{{'form.lookup' | translate}}
-            </button>
-        </div>
     </div>
 
     <ng-container *ngTemplateOutlet="endTemplate?.templateRef; context: model"></ng-container>
 
     <ng-content></ng-content>
-
-
-    <div *ngIf="hasRelationLookup" class="mt-3">
-        <ul class="list-unstyled" cdkDropList (cdkDropListDropped)="moveSelection($event)">
-            <ds-existing-metadata-list-element cdkDrag
-                                               *ngFor="let reorderable of reorderables; trackBy: trackReorderable"
-                                               [reoRel]="reorderable"
-                                               [submissionItem]="item"
-                                               [listId]="listId"
-                                               [metadataFields]="model.metadataFields"
-                                               [relationshipOptions]="model.relationship">
-            </ds-existing-metadata-list-element>
-        </ul>
-    </div>
-</div>
