+++ src/app/shared/form/builder/ds-dynamic-form-ui/ds-dynamic-form.component.ts
@@ -37,5 +37,4 @@ export class DsDynamicFormComponent extends DynamicFormComponent {
   constructor(protected formService: FormBuilderService, protected layoutService: DynamicFormLayoutService) {
     super(formService, layoutService);
   }
-
 }
