+++ src/app/shared/form/form.component.html
@@ -48,7 +48,6 @@
     <div *ngIf="displaySubmit">
       <hr>
       <div class="form-group row">
-
         <div class="col text-right">
           <button type="reset" class="btn btn-default" (click)="reset()">{{'form.cancel' | translate}}</button>
           <button type="submit" class="btn btn-primary" (click)="onSubmit()"
